Wed, Jun 28, 2023, 3:46 PM -04:00

+++ bulbs/bulbs.c
@@ -10,9 +10,9 @@ void print_bulb(int bit);
 int main(void)
 {
     string message = get_string("Message: ");
-    printf("%d\n", strlen(message));
+    printf("%ld\n", strlen(message));
     int bits[8];
-    for (int i = 0; i < strlen(message); i++) //loop through string
+    for (long int i = 0; i < strlen(message); i++) //loop through string
     {
         int j = 0;
         while (message[i] > 0) //dec to binary convertor
